Debugging
Added error handling messages to see where in the program it crashed.

Added a close statement for the sequencer since it was missing.

Also changed the while loop to checking "isRunning" instead of "isOpen".

Partially solved - when the sequence gets to finish it closes the sequencer.

Still a problem trying to push buttons while running, investigating if it has to do with concurrency problem in the JavaFX GUI thread.